@@ -1,14 +1,20 @@
 {
   "name": "react-templates",
-  "version": "0.1.14",
+  "version": "0.3.2",
   "description": "Light weight templates for react -> write html get valid react code",
   "main": "./src/cli.js",
   "bin": {
     "rt": "./bin/rt.js"
   },
   "scripts": {
-    "test": "node test/src/test.js",
-    "test-cov": "istanbul cover test/src/test.js -- --require test/support/env --reporter dot --check-leaks test/ test/acceptance/"
+    "build": "npm run lint && npm run test",
+    "lint": "eslint .",
+    "test": "node test/src/test.js && node test/src/styleTest.js",
+    "test-cov": "istanbul cover test/src/test.js -- --require test/support/env --reporter dot --check-leaks test/ test/acceptance/",
+    "patch": "npm version patch -m\"update version to %s\" && git push && git push --tags",
+    "minor": "npm version minor -m\"update version to %s\" && git push && git push --tags",
+    "major": "npm version major -m\"update version to %s\" && git push && git push --tags",
+    "buildwp": "webpack --config webpack-production.config.js --progress --profile --colors"
   },
   "repository": {
     "type": "git",
@@ -21,28 +27,37 @@
   },
   "homepage": "https://github.com/wix/react-templates",
   "dependencies": {
-    "chalk": "^1.0.0",
-    "cheerio": "^0.18.0",
-    "escodegen": "^1.6.1",
-    "esprima-harmony": "^7001.1.0-dev-harmony-fb",
-    "lodash": "^3.3.0",
-    "optionator": "^0.5.0",
+    "chalk": "^1.1.1",
+    "cheerio": "^0.19.0",
+    "css": "^2.2.1",
+    "escodegen": "^1.7.0",
+    "esprima-fb": "^15001.1001.0-dev-harmony-fb",
+    "lodash": "^3.10.1",
+    "optionator": "^0.6.0",
     "text-table": "^0.2.0"
   },
   "devDependencies": {
-    "brace": "^0.4.1",
-    "brfs": "^1.4.0",
-    "coveralls": "^2.11.2",
+    "babel-eslint": "^4.1.5",
+    "brace": "^0.5.1",
+    "brfs": "^1.4.1",
+    "coveralls": "^2.11.4",
+    "eslint": "^1.10.1",
+    "eslint-config-wix-editor": "0.0.1",
+    "eslint-plugin-lodash3": "^0.2.0",
+    "eslint-plugin-react": "^3.10.0",
+    "eslint-plugin-wix-editor": "^1.0.1",
     "grunt": "^0.4.5",
-    "grunt-browserify": "^3.4.0",
+    "grunt-browserify": "^4.0.0",
     "grunt-contrib-requirejs": "^0.4.4",
-    "grunt-contrib-uglify": "^0.7.0",
+    "grunt-contrib-uglify": "^0.11.0",
     "grunt-contrib-watch": "^0.6.1",
-    "grunt-eslint": "^7.0.0",
+    "grunt-eslint": "^17.3.1",
     "grunt-node-tap": "^0.1.61",
-    "istanbul": "^0.3.6",
-    "react": "^0.12.2",
-    "tape": "^3.5.0"
+    "istanbul": "^0.4.0",
+    "react": "^0.14.2",
+    "react-dom": "^0.14.2",
+    "react-native": "^0.11.4",
+    "tape": "^4.2.1"
   },
   "keywords": [
     "templates",
