@@ -51,8 +51,7 @@ define(['react', 'lodash', 'jquery', './libs/codemirror-4.8/lib/codemirror',
     function completeIfAfterLt(cm) {
         return completeAfter(cm, function () {
             var cur = cm.getCursor();
-            /*eslint new-cap:0*/
-            return cm.getRange(CodeMirror.Pos(cur.line, cur.ch - 1), cur) === '<';
+            return cm.getRange(CodeMirror.Pos(cur.line, cur.ch - 1), cur) === '<'; //eslint-disable-line new-cap
         });
     }
 
@@ -67,13 +66,16 @@ define(['react', 'lodash', 'jquery', './libs/codemirror-4.8/lib/codemirror',
         });
     }
 
-    var editor = React.createClass({
+    return React.createClass({
         displayName: 'CodeMirrorEditor',
         propTypes: {
             id: React.PropTypes.string,
             readOnly: React.PropTypes.bool,
             runMode: React.PropTypes.bool,
-            mode: React.PropTypes.string
+            mode: React.PropTypes.string,
+            value: React.PropTypes.string,
+            valueLink: React.PropTypes.string,
+            onChange: React.PropTypes.func
         },
         getDefaultProps: function () {
             return {
@@ -169,6 +171,4 @@ define(['react', 'lodash', 'jquery', './libs/codemirror-4.8/lib/codemirror',
             this.editor.toTextArea();
         }
     });
-
-    return editor;
 });
