@@ -5,9 +5,9 @@
     include: ['fiddle-main.js'],
     paths: {
         lodash: '//cdnjs.cloudflare.com/ajax/libs/lodash.js/2.4.1/lodash',
-        jquery: '//code.jquery.com/jquery-1.11.0.min',
+        jquery: '//code.jquery.com/jquery-2.1.4.min',
         firebase: '//cdn.firebase.com/js/client/2.0.5/firebase',
-        react: '//fb.me/react-with-addons-0.12.1',
+        react: '//cdnjs.cloudflare.com/ajax/libs/react/0.13.3/react-with-addons',
         //ace: '../ace-builds-1.1.8/src-min/ace',
         fiddle: './fiddle',
         text: 'libs/requirejs-plugins/text',
