@@ -1,15 +1,13 @@
 /**
  * Created by avim on 12/2/2014.
  */
-/*eslint global-strict:0, no-alert:0*/
-/*global alert:true*/
 define(['react', 'firebase', 'lodash', './fiddle.rt', 'jquery'], function (React, Firebase, _, fiddleTemplate, $) {
     'use strict';
 
     function generateRandomId() {
         var uuid = 'xxxxxxxx'.replace(/[xy]/g, function (c) {
             var r = _.random(0, 15);
-            return (c === 'x' ? r : (r & 0x3 | 0x8)).toString(16);
+            return (c === 'x' ? r : r & 0x3 | 0x8).toString(16);
         });
         return uuid;
     }
@@ -37,8 +35,8 @@ define(['react', 'firebase', 'lodash', './fiddle.rt', 'jquery'], function (React
             var firebase = new Firebase('https://reacttemplates.firebaseio-demo.com/');
             firebase.child('fiddles').child(newHash).set(playgroundState, function () {
                 Firebase.goOffline();
-                alert('saved the fiddle, you can share your url');
-            }/*.bind(this)*/);
+                alert('Saved the fiddle, you can share your url'); //eslint-disable-line no-alert
+            });
         },
         clear: function () {
             this.refs.playground.clear();
@@ -59,7 +57,7 @@ define(['react', 'firebase', 'lodash', './fiddle.rt', 'jquery'], function (React
                 $.get('playground/samples/' + name + '.code', null, function (data2/*, textStatus2, jqXHR2*/) {
                     var currentState = {
                         templateHTML: rt,
-                        templateProps: _.template(data2, {name: 'template'})
+                        templateProps: _.template(data2)({name: 'template'})
                     };
                     //this.updateSample(currentState);
                     playground.setState(currentState);
