@@ -1,93 +1,88 @@
-/**
- * Created by avim on 11/9/2014.
- */
 'use strict';
 var cheerio = require('cheerio');
 var _ = require('lodash');
-var esprima = require('esprima-harmony');
+var esprima = require('esprima-fb');
 var escodegen = require('escodegen');
 var reactDOMSupport = require('./reactDOMSupport');
+var reactNativeSupport = require('./reactNativeSupport');
+var reactPropTemplates = require('./reactPropTemplates');
 var stringUtils = require('./stringUtils');
 var rtError = require('./RTCodeError');
+var reactSupport = require('./reactSupport');
+var templates = reactSupport.templates;
+var utils = require('./utils');
+var util = require('util');
+var validateJS = utils.validateJS;
 var RTCodeError = rtError.RTCodeError;
 
 var repeatTemplate = _.template('_.map(<%= collection %>,<%= repeatFunction %>.bind(<%= repeatBinds %>))');
 var ifTemplate = _.template('((<%= condition %>)?(<%= body %>):null)');
 var propsTemplateSimple = _.template('_.assign({}, <%= generatedProps %>, <%= rtProps %>)');
 var propsTemplate = _.template('mergeProps( <%= generatedProps %>, <%= rtProps %>)');
-var propsMergeFunction = 'function mergeProps(inline,external) {\n var res = _.assign({},inline,external)\nif (inline.hasOwnProperty(\'style\')) {\n res.style = _.defaults(res.style, inline.style);\n}\n if (inline.hasOwnProperty(\'className\') && external.hasOwnProperty(\'className\')) {\n res.className = external.className + \' \' + inline.className;\n} return res;\n}\n';
-var classSetTemplate = _.template('React.addons.classSet(<%= classSet %>)');
+
+var propsMergeFunction = [
+    'function mergeProps(inline,external) {',
+    ' var res = _.assign({},inline,external)',
+    'if (inline.hasOwnProperty(\'style\')) {',
+    ' res.style = _.defaults(res.style, inline.style);',
+    '}',
+    ' if (inline.hasOwnProperty(\'className\') && external.hasOwnProperty(\'className\')) {',
+    ' res.className = external.className + \' \' + inline.className;',
+    '} return res;',
+    '}',
+    ''
+].join('\n');
+
+var classSetTemplate = _.template('_.keys(_.pick(<%= classSet %>, _.identity)).join(" ")');
 var simpleTagTemplate = _.template('<%= name %>(<%= props %><%= children %>)');
-var tagTemplate = _.template('<%= name %>.apply(this,_.flatten([<%= props %><%= children %>]))');
+var tagTemplate = _.template('<%= name %>.apply(this, [<%= props %><%= children %>])');
 var simpleTagTemplateCreateElement = _.template('React.createElement(<%= name %>,<%= props %><%= children %>)');
-var tagTemplateCreateElement = _.template('React.createElement.apply(this,_.flatten([<%= name %>,<%= props %><%= children %>]))');
+var tagTemplateCreateElement = _.template('React.createElement.apply(this, [<%= name %>,<%= props %><%= children %>])');
 var commentTemplate = _.template(' /* <%= data %> */ ');
 
-var templateAMDTemplate = _.template("define(<%= name ? '\"'+name + '\", ' : '' %>[<%= requirePaths %>], function (<%= requireNames %>) {\n'use strict';\n <%= injectedFunctions %>\nreturn function(){ return <%= body %>};\n});");
-var templateCommonJSTemplate = _.template("'use strict';\n<%= vars %>\n\n<%= injectedFunctions %>\nmodule.exports = function(){ return <%= body %>};\n");
-var templateES6Template = _.template('<%= vars %>\n\n<%= injectedFunctions %>\nexport default function(){ return <%= body %>}\n');
-var templatePJSTemplate = _.template('var <%= name %> = function () {\n' +
-                                '<%= injectedFunctions %>\n' +
-                                'return <%= body %>\n' +
-                                '};\n');
-var templateTypescriptTemplate = _.template('<%= vars %>\n\n<%= injectedFunctions %>\nvar fn = function() { return <%= body %> };\nexport = fn\n');
+var repeatAttr = 'rt-repeat';
+var ifAttr = 'rt-if';
+var classSetAttr = 'rt-class';
+var classAttr = 'class';
+var scopeAttr = 'rt-scope';
+var propsAttr = 'rt-props';
+var templateNode = 'rt-template';
+var skippedProps = ["data-bind"];
 
-var templates = {
-    amd: templateAMDTemplate,
-    commonjs: templateCommonJSTemplate,
-    typescript: templateTypescriptTemplate,
-    es6: templateES6Template,
-    none: templatePJSTemplate
-};
+function getOptions(options) {
+    options = options || {};
+    var defaultOptions = {
+        modules: options.native ? 'commonjs' : 'amd',
+        version: false,
+        force: false,
+        format: 'stylish',
+        targetVersion: reactDOMSupport.default,
+        reactImportPath: reactImport(options),
+        lodashImportPath: 'lodash',
+        native: false,
+        nativeTargetVersion: reactNativeSupport.default,
+        flow: options.flow,
+        enableSkipRequire: 'false'
+    };
 
-var htmlSelfClosingTags = ['area', 'base', 'br', 'col', 'command', 'embed', 'hr', 'img', 'input', 'keygen', 'link', 'meta', 'param', 'source', 'track', 'wbr'];
+    var finalOptions = _.defaults({}, options, defaultOptions);
 
-var templateProp = 'rt-repeat';
-var ifProp = 'rt-if';
-var classSetProp = 'rt-class';
-var scopeProp = 'rt-scope';
-var propsProp = 'rt-props';
-var skippedProps = ["data-bind"]
-var defaultOptions = {modules: 'amd', version: false, force: false, format: 'stylish', targetVersion: '0.12.2'};
+    var defaultPropTemplates = finalOptions.native ?
+        reactPropTemplates.native[finalOptions.nativeTargetVersion] :
+        reactPropTemplates.dom[finalOptions.targetVersion];
 
-/**
- * @param {Context} context
- * @return {boolean}
- */
-function shouldUseCreateElement(context) {
-    switch (context.options.targetVersion) {
-        case '0.11.2':
-        case '0.11.1':
-        case '0.11.0':
-        case '0.10.0':
-            return false;
-        default:
-            return true;
-    }
+    finalOptions.propTemplates = _.defaults({}, options.propTemplates, defaultPropTemplates);
+    return finalOptions;
 }
 
-var reactSupportedAttributes = ['accept', 'acceptCharset', 'accessKey', 'action', 'allowFullScreen', 'allowTransparency', 'alt', 'async', 'autoComplete', 'autoPlay', 'cellPadding', 'cellSpacing', 'charSet', 'checked', 'classID', 'className', 'cols', 'colSpan', 'content', 'contentEditable', 'contextMenu', 'controls', 'coords', 'crossOrigin', 'data', 'dateTime', 'defer', 'dir', 'disabled', 'download', 'draggable', 'encType', 'form', 'formNoValidate', 'frameBorder', 'height', 'hidden', 'href', 'hrefLang', 'htmlFor', 'httpEquiv', 'icon', 'id', 'label', 'lang', 'list', 'loop', 'manifest', 'max', 'maxLength', 'media', 'mediaGroup', 'method', 'min', 'multiple', 'muted', 'name', 'noValidate', 'open', 'pattern', 'placeholder', 'poster', 'preload', 'radioGroup', 'readOnly', 'rel', 'required', 'role', 'rows', 'rowSpan', 'sandbox', 'scope', 'scrolling', 'seamless', 'selected', 'shape', 'size', 'sizes', 'span', 'spellCheck', 'src', 'srcDoc', 'srcSet', 'start', 'step', 'style', 'tabIndex', 'target', 'title', 'type', 'useMap', 'value', 'width', 'wmode'];
-var attributesMapping = {'class': 'className', 'rt-class': 'className'};
-_.forEach(reactSupportedAttributes, function (attributeReactName) {
-    if (attributeReactName !== attributeReactName.toLowerCase()) {
-        attributesMapping[attributeReactName.toLowerCase()] = attributeReactName;
+function reactImport(options) {
+    if (options.native) {
+        return 'react-native';
     }
-});
-
-/**
- * @param children
- * @return {string}
- */
-function concatChildren(children) {
-    var res = '';
-    _.forEach(children, function (child) {
-        if (child && child.indexOf(' /*') !== 0 ) {
-            res += ',' + child;
-        } else {
-            res += child;
-        }
-    }, this);
-    return res;
+    if (options.targetVersion === '0.14.0') {
+        return 'react';
+    }
+    return 'react/addons';
 }
 
 /**
@@ -122,13 +117,13 @@ function convertText(node, context, txt) {
         for (end = start + 1; end < txt.length && curlyCounter > 0; end++) {
             curlyCounter += curlyMap[txt.charAt(end)] || 0;
         }
-        if (curlyCounter !== 0) {
-            throw RTCodeError.build("Failed to parse text '" + txt + "'", context, node);
-        } else {
+        if (curlyCounter === 0) {
             var needsParens = start !== 0 || end !== txt.length - 1;
             res += (first ? '' : concatChar) + (needsParens ? '(' : '') + txt.substr(start + 1, end - start - 2) + (needsParens ? ')' : '');
             first = false;
             txt = txt.substr(end);
+        } else {
+            throw RTCodeError.buildFormat(context, node, "Failed to parse text '%s'", txt);
         }
     }
     if (txt) {
@@ -138,17 +133,7 @@ function convertText(node, context, txt) {
         res = 'true';
     }
 
-    //validateJS(res, node, context);
     return res;
-}
-
-/**
- * @param {string} txt
- * @return {boolean}
- */
-function isStringOnlyCode(txt) {
-    txt = txt.trim();
-    return txt.length && txt.charAt(0) === '{' && txt.charAt(txt.length - 1) === '}';
 }
 
 /**
@@ -161,10 +146,57 @@ function isStringOnlyCode(txt) {
 function generateInjectedFunc(context, namePrefix, body, params) {
     params = params || context.boundParams;
     var generatedFuncName = namePrefix.replace(',', '') + (context.injectedFunctions.length + 1);
-    var funcText = 'function ' + generatedFuncName + '(' + params.join(',');
-    funcText += ') {\n' + body + '\n}\n';
+    var funcText = util.format('function %s(%s) {\n%s\n}\n', generatedFuncName, params.join(','), body);
     context.injectedFunctions.push(funcText);
     return generatedFuncName;
+}
+
+function generateTemplateProps(node, context) {
+    var propTemplateDefinition = context.options.propTemplates[node.name];
+    var propertiesTemplates = _(node.children)
+        .map(function (child, index) {
+            var templateProp = null;
+            if (child.name === templateNode) { // Generic explicit template tag
+                if (!_.has(child.attribs, 'prop')) {
+                    throw RTCodeError.build('rt-template must have a prop attribute', context, child);
+                }
+
+                var childTemplate = _.find(context.options.propTemplates, {prop: child.attribs.prop}) || {arguments: []};
+                templateProp = {
+                    prop: child.attribs.prop,
+                    arguments: (child.attribs.arguments ? child.attribs.arguments.split(',') : childTemplate.arguments) || []
+                };
+            } else if (propTemplateDefinition && propTemplateDefinition[child.name]) { // Implicit child template from configuration
+                templateProp = {
+                    prop: propTemplateDefinition[child.name].prop,
+                    arguments: child.attribs.arguments ? child.attribs.arguments.split(',') : propTemplateDefinition[child.name].arguments
+                };
+            }
+
+            if (templateProp) {
+                _.assign(templateProp, {childIndex: index, content: _.find(child.children, {type: 'tag'})});
+            }
+
+            return templateProp;
+        })
+        .compact()
+        .value();
+
+    return _.transform(propertiesTemplates, function (props, templateProp) {
+        var functionParams = _.values(context.boundParams).concat(templateProp.arguments);
+
+        var oldBoundParams = context.boundParams;
+        context.boundParams = context.boundParams.concat(templateProp.arguments);
+
+        var functionBody = 'return ' + convertHtmlToReact(templateProp.content, context);
+        context.boundParams = oldBoundParams;
+
+        var generatedFuncName = generateInjectedFunc(context, templateProp.prop, functionBody, functionParams);
+        props[templateProp.prop] = genBind(generatedFuncName, _.values(context.boundParams));
+
+        // Remove the template child from the children definition.
+        node.children.splice(templateProp.childIndex, 1);
+    }, {});
 }
 
 /**
@@ -173,55 +205,74 @@ function generateInjectedFunc(context, namePrefix, body, params) {
  * @return {string}
  */
 function generateProps(node, context) {
-//    console.log(node);
     var props = {};
     _.forOwn(node.attribs, function (val, key) {
-        var propKey = attributesMapping[key.toLowerCase()] || key;
-        if (props.hasOwnProperty(propKey)) {
-            throw RTCodeError.build('duplicate definition of ' + propKey + ' ' + JSON.stringify(node.attribs), context, node);
+        var propKey = reactSupport.attributesMapping[key.toLowerCase()] || key;
+        if (props.hasOwnProperty(propKey) && propKey !== reactSupport.classNameProp) {
+            throw RTCodeError.buildFormat(context, node, 'duplicate definition of %s %s', propKey, JSON.stringify(node.attribs));
         }
         if (_.contains(skippedProps, key)) {
           // if the prop can be skipped
-        } else if (key.indexOf('on') === 0 && !isStringOnlyCode(val)) {
-            var funcParts = val.split('=>');
-            if (funcParts.length !== 2) {
-                throw RTCodeError.build("when using 'on' events, use lambda '(p1,p2)=>body' notation or use {} to return a callback function. error: [" + key + "='" + val + "']", context, node);
+        } else if (key.indexOf('on') === 0 && !utils.isStringOnlyCode(val)) {
+            props[propKey] = handleEventHandler(val, context, node, key);
+        } else if (key === 'style' && !utils.isStringOnlyCode(val)) {
+            props[propKey] = handleStyleProp(val, node, context);
+        } else if (propKey === reactSupport.classNameProp) {
+            // Processing for both class and rt-class conveniently return strings that
+            // represent JS expressions, each evaluating to a space-separated set of class names.
+            // We can just join them with another space here.
+            var existing = props[propKey] ? props[propKey] + ' + " " + ' : '';
+            if (key === classSetAttr) {
+                props[propKey] = existing + classSetTemplate({classSet: val});
+            } else if (key === classAttr || key === reactSupport.classNameProp) {
+                props[propKey] = existing + convertText(node, context, val.trim());
             }
-            var evtParams = funcParts[0].replace('(', '').replace(')', '').trim();
-            var funcBody = funcParts[1].trim();
-            var params = context.boundParams;
-            if (evtParams.trim() !== '') {
-                params = params.concat([evtParams.trim()]);
-            }
-            var generatedFuncName = generateInjectedFunc(context, key, funcBody, params);
-            props[propKey] = generatedFuncName + '.bind(' + (['this'].concat(context.boundParams)).join(',') + ')';
-        } else if (key === 'style' && !isStringOnlyCode(val)) {
-            var styleParts = val.trim().split(';');
-            styleParts = _.compact(_.map(styleParts, function (str) {
-                str = str.trim();
-                if (!str || str.indexOf(':') === -1) {
-                    return null;
-                }
-                var res = str.split(':');
-                res[0] = res[0].trim();
-                res[1] = res.slice(1).join(':').trim();
-                return res;
-            }));
-            var styleArray = [];
-            _.forEach(styleParts, function (stylePart) {
-                styleArray.push(stringUtils.convertToCamelCase(stylePart[0]) + ' : ' + convertText(node, context, stylePart[1].trim()));
-            });
-            props[propKey] = '{' + styleArray.join(',') + '}';
-        } else if (key === classSetProp) {
-            props[propKey] = classSetTemplate({classSet: val});
         } else if (key.indexOf('rt-') !== 0) {
             props[propKey] = convertText(node, context, val.trim());
         }
     });
+    _.assign(props, generateTemplateProps(node, context));
 
     return '{' + _.map(props, function (val, key) {
         return JSON.stringify(key) + ' : ' + val;
     }).join(',') + '}';
+}
+
+function handleEventHandler(val, context, node, key) {
+    var funcParts = val.split('=>');
+    if (funcParts.length !== 2) {
+        throw RTCodeError.buildFormat(context, node, "when using 'on' events, use lambda '(p1,p2)=>body' notation or use {} to return a callback function. error: [%s='%s']", key, val);
+    }
+    var evtParams = funcParts[0].replace('(', '').replace(')', '').trim();
+    var funcBody = funcParts[1].trim();
+    var params = context.boundParams;
+    if (evtParams.trim() !== '') {
+        params = params.concat([evtParams.trim()]);
+    }
+    var generatedFuncName = generateInjectedFunc(context, key, funcBody, params);
+    return genBind(generatedFuncName, context.boundParams);
+}
+
+function genBind(func, args) {
+    return util.format('%s.bind(%s)', func, (['this'].concat(args)).join(','));
+}
+
+function handleStyleProp(val, node, context) {
+    var styleParts = val.trim().split(';');
+    styleParts = _.compact(_.map(styleParts, function (str) {
+        str = str.trim();
+        if (!str || str.indexOf(':') === -1) {
+            return null;
+        }
+        var res = str.split(':');
+        res[0] = res[0].trim();
+        res[1] = res.slice(1).join(':').trim();
+        return res;
+    }));
+    var styleArray = _.map(styleParts, function (stylePart) {
+        return stringUtils.convertToCamelCase(stylePart[0]) + ' : ' + convertText(node, context, stylePart[1].trim());
+    });
+    return '{' + styleArray.join(',') + '}';
 }
 
 /**
@@ -230,8 +281,11 @@ function generateProps(node, context) {
  * @return {string}
  */
 function convertTagNameToConstructor(tagName, context) {
-    var isHtmlTag = _.contains(reactDOMSupport[context.options.targetVersion], tagName);
-    if (shouldUseCreateElement(context)) {
+    if (context.options.native) {
+        return _.includes(reactNativeSupport[context.options.nativeTargetVersion], tagName) ? 'React.' + tagName : tagName;
+    }
+    var isHtmlTag = _.includes(reactDOMSupport[context.options.targetVersion], tagName);
+    if (reactSupport.shouldUseCreateElement(context)) {
         isHtmlTag = isHtmlTag || tagName.match(/^\w+(-\w+)$/);
         return isHtmlTag ? "'" + tagName + "'" : tagName;
     }
@@ -257,8 +311,8 @@ function defaultContext(html, options) {
  * @return {boolean}
  */
 function hasNonSimpleChildren(node) {
-    return _.any(node.children, function (child) {
-        return child.type === 'tag' && child.attribs[templateProp];
+    return _.some(node.children, function (child) {
+        return child.type === 'tag' && child.attribs[repeatAttr];
     });
 }
 
@@ -277,83 +331,104 @@ function convertHtmlToReact(node, context) {
         };
 
         var data = {name: convertTagNameToConstructor(node.name, context)};
-        if (node.attribs[scopeProp]) {
-            data.scopeMapping = {};
-            data.scopeName = '';
-            _.each(context.boundParams, function (boundParam) {
-                data.scopeMapping[boundParam] = boundParam;
-            });
-            _.each(node.attribs[scopeProp].split(';'), function (scopePart) {
-                var scopeSubParts = scopePart.split(' as ');
-                if (scopeSubParts.length < 2) {
-                    throw RTCodeError.build("invalid scope part '" + scopePart + "'", context, node);
-                }
-                var scopeName = scopeSubParts[1].trim();
-                validateJS(scopeName, node, context);
-                stringUtils.addIfNotThere(context.boundParams, scopeName);
-                data.scopeName += stringUtils.capitalize(scopeName);
-                data.scopeMapping[scopeName] = scopeSubParts[0].trim();
-                validateJS(data.scopeMapping[scopeName], node, context);
-            });
+        
+        if (context.options.enableSkipRequire == 'true' && node.attribs.skip == 'true') {
+          return 'null';
         }
 
-        if (node.attribs[templateProp]) {
-            var arr = node.attribs[templateProp].split(' in ');
+        // Order matters. We need to add the item and itemIndex to context.boundParams before
+        // the rt-scope directive is processed, lest they are not passed to the child scopes
+        if (node.attribs[repeatAttr]) {
+            var arr = node.attribs[repeatAttr].split(' in ');
             if (arr.length !== 2) {
-                throw RTCodeError.build("rt-repeat invalid 'in' expression '" + node.attribs[templateProp] + "'", context, node);
+                throw RTCodeError.buildFormat(context, node, "rt-repeat invalid 'in' expression '%s'", node.attribs[repeatAttr]);
             }
             data.item = arr[0].trim();
             data.collection = arr[1].trim();
             validateJS(data.item, node, context);
             validateJS(data.collection, node, context);
-            stringUtils.addIfNotThere(context.boundParams, data.item);
-            stringUtils.addIfNotThere(context.boundParams, data.item + 'Index');
+            stringUtils.addIfMissing(context.boundParams, data.item);
+            stringUtils.addIfMissing(context.boundParams, data.item + 'Index');
         }
+
+        if (node.attribs[scopeAttr]) {
+            data.innerScope = {
+                scopeName: '',
+                innerMapping: {},
+                outerMapping: {}
+            };
+
+            data.innerScope.outerMapping = _.zipObject(context.boundParams, context.boundParams);
+
+            _(node.attribs[scopeAttr]).split(';').invoke('trim').compact().forEach( function (scopePart) {
+                var scopeSubParts = _(scopePart).split(' as ').invoke('trim').value();
+                if (scopeSubParts.length < 2) {
+                    throw RTCodeError.buildFormat(context, node, "invalid scope part '%s'", scopePart);
+                }
+                var alias = scopeSubParts[1];
+                var value = scopeSubParts[0];
+                validateJS(alias, node, context);
+
+                // this adds both parameters to the list of parameters passed further down
+                // the scope chain, as well as variables that are locally bound before any
+                // function call, as with the ones we generate for rt-scope.
+                stringUtils.addIfMissing(context.boundParams, alias);
+
+                data.innerScope.scopeName += stringUtils.capitalize(alias);
+                data.innerScope.innerMapping[alias] = 'var ' + alias + ' = ' + value + ';';
+                validateJS(data.innerScope.innerMapping[alias], node, context);
+            }).value();
+        }
+
         data.props = generateProps(node, context);
-        if (node.attribs[propsProp]) {
+        if (node.attribs[propsAttr]) {
             if (data.props === '{}') {
-                data.props = node.attribs[propsProp];
+                data.props = node.attribs[propsAttr];
             } else if (!node.attribs.style && !node.attribs.class) {
-                data.props = propsTemplateSimple({generatedProps: data.props, rtProps: node.attribs[propsProp]});
+                data.props = propsTemplateSimple({generatedProps: data.props, rtProps: node.attribs[propsAttr]});
             } else {
-                data.props = propsTemplate({generatedProps: data.props, rtProps: node.attribs[propsProp]});
-                if (!_.contains(context.injectedFunctions,propsMergeFunction)) {
+                data.props = propsTemplate({generatedProps: data.props, rtProps: node.attribs[propsAttr]});
+                if (!_.includes(context.injectedFunctions, propsMergeFunction)) {
                     context.injectedFunctions.push(propsMergeFunction);
                 }
             }
         }
-        if (node.attribs[ifProp]) {
-            data.condition = node.attribs[ifProp].trim();
+        if (node.attribs[ifAttr]) {
+            data.condition = node.attribs[ifAttr].trim();
         }
-        data.children = node.children ? concatChildren(_.map(node.children, function (child) {
+        data.children = utils.concatChildren(_.map(node.children, function (child) {
             var code = convertHtmlToReact(child, context);
             validateJS(code, child, context);
             return code;
-        })) : '';
+        }));
 
         if (hasNonSimpleChildren(node)) {
-            data.body = shouldUseCreateElement(context) ? tagTemplateCreateElement(data) : tagTemplate(data);
+            data.body = reactSupport.shouldUseCreateElement(context) ? tagTemplateCreateElement(data) : tagTemplate(data);
         } else {
-            data.body = shouldUseCreateElement(context) ? simpleTagTemplateCreateElement(data) : simpleTagTemplate(data);
+            data.body = reactSupport.shouldUseCreateElement(context) ? simpleTagTemplateCreateElement(data) : simpleTagTemplate(data);
         }
 
-        if (node.attribs[templateProp]) {
+        if (node.attribs[scopeAttr]) {
+            var functionBody = _.values(data.innerScope.innerMapping).join('\n') + 'return ' + data.body;
+            var generatedFuncName = generateInjectedFunc(context, 'scope' + data.innerScope.scopeName, functionBody, _.keys(data.innerScope.outerMapping));
+            data.body = util.format('%s.apply(this, [%s])', generatedFuncName, _.values(data.innerScope.outerMapping).join(','));
+        }
+
+        // Order matters here. Each rt-repeat iteration wraps over the rt-scope, so
+        // the scope variables are evaluated in context of the current iteration.
+        if (node.attribs[repeatAttr]) {
             data.repeatFunction = generateInjectedFunc(context, 'repeat' + stringUtils.capitalize(data.item), 'return ' + data.body);
             data.repeatBinds = ['this'].concat(_.reject(context.boundParams, function (param) {
-                return (param === data.item || param === data.item + 'Index');
+                return param === data.item || param === data.item + 'Index' || data.innerScope && param in data.innerScope.innerMapping;
             }));
             data.body = repeatTemplate(data);
         }
-        if (node.attribs[ifProp]) {
+        if (node.attribs[ifAttr]) {
             data.body = ifTemplate(data);
-        }
-        if (node.attribs[scopeProp]) {
-            var generatedFuncName = generateInjectedFunc(context, 'scope' + data.scopeName, 'return ' + data.body, _.keys(data.scopeMapping));
-            data.body = generatedFuncName + '.apply(this, [' + _.values(data.scopeMapping).join(',') + '])';
         }
         return data.body;
     } else if (node.type === 'comment') {
-        return (commentTemplate(node));
+        return commentTemplate(node);
     } else if (node.type === 'text') {
         if (node.data.trim()) {
             return convertText(node, context, node.data);
@@ -362,12 +437,6 @@ function convertHtmlToReact(node, context) {
     }
 }
 
-//function removeDocType(html) {
-//  html = html.replace(/^\s*\<\!doctype\s+rt\s*>/mi, function () {
-//    return '';
-//  });
-//  return html;
-//}
 /**
  * @param node
  * @return {boolean}
@@ -381,7 +450,7 @@ function handleSelfClosingHtmlTags(nodes) {
         .map(function (node) {
             var externalNodes = [];
             node.children = handleSelfClosingHtmlTags(node.children);
-            if (node.type === 'tag' && _.contains(htmlSelfClosingTags, node.name)) {
+            if (node.type === 'tag' && _.includes(reactSupport.htmlSelfClosingTags, node.name)) {
                 externalNodes = _.filter(node.children, isTag);
                 _.forEach(externalNodes, function (child) {
                     child.parent = node;
@@ -394,17 +463,30 @@ function handleSelfClosingHtmlTags(nodes) {
         .value();
 }
 
+function convertTemplateToReact(html, options) {
+    var context = require('./context');
+    return convertRT(html, context, options);
+}
+
 /**
  * @param {string} html
+ * @param {CONTEXT} reportContext
  * @param {{modules:string,defines:*}?} options
  * @return {string}
  */
-function convertTemplateToReact(html, options) {
+function convertRT(html, reportContext, options) {
     var rootNode = cheerio.load(html, {lowerCaseTags: false, lowerCaseAttributeNames: false, xmlMode: true, withStartIndices: true});
-    options = _.defaults({}, options, defaultOptions);
-    var defines = options.defines ? options.defines : {'react/addons': 'React', lodash: '_'};
+    options = getOptions(options);
+
+    var defaultDefines = {};
+    defaultDefines[options.reactImportPath] = 'React';
+    defaultDefines[options.lodashImportPath] = '_';
+
+    var defines = options.defines ? _.clone(options.defines) : defaultDefines;
+
     var context = defaultContext(html, options);
-    var rootTags = _.filter(rootNode.root()[0].children, {type: 'tag'});
+    utils.validate(options, context, reportContext, rootNode.root()[0]);
+    var rootTags = _.filter(rootNode.root()[0].children, isTag);
     rootTags = handleSelfClosingHtmlTags(rootTags);
     if (!rootTags || rootTags.length === 0) {
         throw new RTCodeError('Document should have a root element');
@@ -417,11 +499,7 @@ function convertTemplateToReact(html, options) {
             } else if (tag.children.length) {
                 throw RTCodeError.build('rt-require may have no children', context, tag);
             }
-            //if (options.modules === 'typescript') {
-            //    defines['./' + tag.attribs.dependency] = tag.attribs.as;
-            //} else {
             defines[tag.attribs.dependency] = tag.attribs.as;
-            //}
         } else if (firstTag === null) {
             firstTag = tag;
         } else {
@@ -432,28 +510,57 @@ function convertTemplateToReact(html, options) {
         throw RTCodeError.build('Document should have a single root element', context, rootNode.root()[0]);
     }
     var body = convertHtmlToReact(firstTag, context);
-    var requirePaths = _(defines).keys().map(function (reqName) { return '"' + reqName + '"'; }).value().join(',');
-    var requireVars = _(defines).values().value().join(',');
-    var vars;
+    var requirePaths = _(defines)
+        .keys()
+        .map(function (reqName) { return '"' + reqName + '"'; })
+        .join(',');
+    var requireVars = _.values(defines).join(',');
+    var buildImport;
     if (options.modules === 'typescript') {
-        vars = _(defines).map(function (reqVar, reqPath) { return 'import ' + reqVar + " = require('" + reqPath + "');"; }).join('\n');
+        buildImport = function (reqVar, reqPath) {
+            return util.format("import %s = require('%s');", reqVar, reqPath);
+        };
     } else if (options.modules === 'es6') {
-        vars = _(defines).map(function (reqVar, reqPath) { return 'import {' + reqVar + "} from '" + reqPath + "';"; }).join('\n');
+        buildImport = function (reqVar, reqPath) {
+            return util.format("import %s from '%s';", reqVar, reqPath);
+        };
     } else {
-        vars = _(defines).map(function (reqVar, reqPath) { return 'var ' + reqVar + " = require('" + reqPath + "');"; }).join('\n');
+        buildImport = function (reqVar, reqPath) {
+            return util.format("var %s = require('%s');", reqVar, reqPath);
+        };
+    }
+    var vars = _(defines).map(buildImport).join('\n');
+
+    if (options.flow) {
+        vars = '/* @flow */\n' + vars;
     }
     var data = {body: body, injectedFunctions: '', requireNames: requireVars, requirePaths: requirePaths, vars: vars, name: options.name};
     data.injectedFunctions = context.injectedFunctions.join('\n');
     var code = generate(data, options);
-    if (options.modules !== 'typescript') {
-        try {
-            var tree = esprima.parse(code, {range: true, tokens: true, comment: true});
-            tree = escodegen.attachComments(tree, tree.comments, tree.tokens);
-            code = escodegen.generate(tree, {comment: true});
-        } catch (e) {
-            throw new RTCodeError(e.message, e.index, -1);
-        }
+    if (options.modules !== 'typescript' && options.modules !== 'jsrt') {
+        code = parseJS(code);
     }
+    return code;
+}
+
+function parseJS(code) {
+    try {
+        var tree = esprima.parse(code, {range: true, tokens: true, comment: true, sourceType: 'module'});
+        tree = escodegen.attachComments(tree, tree.comments, tree.tokens);
+        return escodegen.generate(tree, {comment: true});
+    } catch (e) {
+        throw new RTCodeError(e.message, e.index, -1);
+    }
+}
+
+function convertJSRTToJS(text, reportContext, options) {
+    options = getOptions(options);
+    options.modules = 'jsrt';
+    var templateMatcherJSRT = /<template>([^]*?)<\/template>/gm;
+    var code = text.replace(templateMatcherJSRT, function (template, html) {
+        return convertRT(html, reportContext, options).replace(/;$/, '');
+    });
+    code = parseJS(code);
     return code;
 }
 
@@ -462,31 +569,12 @@ function generate(data, options) {
     return template(data);
 }
 
-/**
- * @param {string} code
- * @param node
- * @param {Context} context
- */
-function validateJS(code, node, context) {
-    try {
-        esprima.parse(code);
-    } catch (e) {
-        throw RTCodeError.build(e.description, context, node);
-    }
-}
-
-/**
- * @param {string} name
- * @return {string}
- */
-function normalizeName(name) {
-    return name.replace(/-/g, '_');
-}
-
 module.exports = {
     convertTemplateToReact: convertTemplateToReact,
+    convertRT: convertRT,
+    convertJSRTToJS: convertJSRTToJS,
     RTCodeError: RTCodeError,
-    normalizeName: normalizeName,
+    normalizeName: utils.normalizeName,
     _test: {
         convertText: convertText
     }
