@@ -1,5 +1,5 @@
-import { React } from 'react/addons';
-import { _ } from 'lodash';
+import React from 'react/addons';
+import _ from 'lodash';
 export default function () {
     return React.createElement('div', {});
 };
